Reorganize capabilities packages and add manifests
Moves capabilities related packages to `pkg/v1/sdk` and adds manifests
for deploying capabilities controller.
