+++ pkg/v1/sdk/capabilities/config/rbac.yaml
@@ -1,0 +1,82 @@
+apiVersion: v1
+kind: ServiceAccount
+metadata:
+  labels:
+    app: tanzu-capabilities-manager
+  name: tanzu-capabilities-manager-sa
+  namespace: tkg-system
+---
+apiVersion: rbac.authorization.k8s.io/v1
+kind: ClusterRole
+metadata:
+  name: tanzu-capabilities-manager-clusterrole
+rules:
+  - apiGroups:
+      - run.tanzu.vmware.com
+    resources:
+      - capabilities
+    verbs:
+      - create
+      - delete
+      - get
+      - list
+      - patch
+      - update
+      - watch
+  - apiGroups:
+      - run.tanzu.vmware.com
+    resources:
+      - capabilities/status
+    verbs:
+      - get
+      - patch
+      - update
+  - apiGroups:
+      - run.tanzu.vmware.com
+    resources:
+      - tanzukubernetesreleases
+      - tanzukubernetesreleases/status
+    verbs:
+      - get
+      - list
+      - watch
+  - apiGroups:
+      - run.tanzu.vmware.com
+    resources:
+      - tanzukubernetesclusters
+      - tanzukubernetesclusters/status
+    verbs:
+      - get
+      - list
+      - watch
+  - apiGroups:
+      - clusterctl.cluster.x-k8s.io
+    resources:
+      - providers
+      - providers/status
+    verbs:
+      - get
+      - list
+      - watch
+  - apiGroups:
+      - ""
+    resources:
+      - configmaps
+      - nodes
+    verbs:
+      - get
+      - list
+      - watch
+---
+apiVersion: rbac.authorization.k8s.io/v1
+kind: ClusterRoleBinding
+metadata:
+  name: tanzu-capabilities-manager-clusterrolebinding
+roleRef:
+  apiGroup: rbac.authorization.k8s.io
+  kind: ClusterRole
+  name: tanzu-capabilities-manager-clusterrole
+subjects:
+  - kind: ServiceAccount
+    name: tanzu-capabilities-manager-sa
+    namespace: tkg-system
