+++ pkg/v1/sdk/capabilities/config/tanzu-capabilities-manager.yaml
@@ -1,0 +1,33 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  labels:
+    app: tanzu-capabilities-manager
+  name: tanzu-capabilities-controller-manager
+  namespace: tkg-system
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app: tanzu-capabilities-manager
+  template:
+    metadata:
+      labels:
+        app: tanzu-capabilities-manager
+    spec:
+      containers:
+        - image:  capabilities-controller-manager:latest
+          imagePullPolicy: IfNotPresent
+          name: manager
+          resources:
+            limits:
+              cpu: 100m
+              memory: 30Mi
+            requests:
+              cpu: 100m
+              memory: 20Mi
+      serviceAccount: tanzu-capabilities-manager-sa
+      terminationGracePeriodSeconds: 10
+      tolerations:
+        - effect: NoSchedule
+          key: node-role.kubernetes.io/master
