+++ pkg/v1/sdk/capabilities/controllers/doc.go
@@ -1,5 +1,5 @@
 // Copyright 2021 VMware, Inc. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-// Package controllers has controllers for run.tanzu.vmware.com APIs.
+// Package controllers has controllers for Capability API.
 package controllers
