+++ pkg/v1/sdk/capabilities/discovery/README.md
