+++ pkg/v1/sdk/capabilities/discovery/cluster_object.go
