+++ pkg/v1/sdk/capabilities/discovery/cluster_test.go
