+++ pkg/v1/sdk/capabilities/discovery/discovery.go
