+++ pkg/v1/sdk/capabilities/discovery/doc.go
