+++ pkg/v1/sdk/capabilities/discovery/fake.go
