+++ pkg/v1/sdk/capabilities/discovery/generate.go
