+++ pkg/v1/sdk/capabilities/discovery/tkg/capabilities_test.go
