+++ pkg/v1/sdk/capabilities/discovery/tkg/doc.go
