+++ pkg/v1/sdk/capabilities/discovery/tkg/fake.go
@@ -9,7 +9,7 @@ import (
 	ctrlfake "sigs.k8s.io/controller-runtime/pkg/client/fake" //nolint:staticcheck
 
 	runv1alpha1 "github.com/vmware-tanzu/tanzu-framework/apis/run/v1alpha1"
-	"github.com/vmware-tanzu/tanzu-framework/pkg/v1/discovery"
+	"github.com/vmware-tanzu/tanzu-framework/pkg/v1/sdk/capabilities/discovery"
 )
 
 var (
