+++ pkg/v1/sdk/capabilities/discovery/tkg/resource_test.go
